chore: update dependencies for "consistent-as-needed" (#38)
As noted in this comment:

https://github.com/liferay/eslint-config-liferay/pull/39#issuecomment-475596137

we need the latest Prettier (1.17.0) in order to make use of the
"consistent-as-needed" rule, and in our Prettier config, we'll need to
use `"quote-props": "consistent"`.

So, I updated everything using `yarn upgrade-interactive` and added the
necessary rule to the ".prettierrc.json" file; other projects will need
to do the same.